Declare armel and armhf to dpkg as foreign architectures
Any non-native sdk (e.g., for an armhf device) needs to satisfy
build-dependencies using packages from the device's architecture. dpkg
allows packages of a foreign architecture to be installed, but those
architectures must be made known with the command 'dpkg
--add-architecture $arch' prior to downloading APT source metadata with
'apt-get update'.

We used to do this manually, but with this change, we go ahead and add
all Pragmatux device architectures when the apt sources are installed.

The only downside is that 'apt-get update' on the workstation will now
pull metadata from unnecessary architectures. (That sometimes leads to
errors if the upsteam Debian mirror at which the workstation is pointing
doesn't have all architectures. Use a mirror which does.)