Don't inadvertently draw in amd64 *device* packages
amd64 was added so that library packages from amd64 targets could be
drawn in for cross-building (not cross-arch but cross-OS, i.e. a
different set of library versions than the host OS). This creates a
problem because all device packages suddenly become available to the
host for installation. If the device repo has a newer udev, for example,
then the host's udev will get clobbered.